Contracts: Create getRecipientClaimData function and move it to utils/maci module
