+++ contracts/scripts/claim.ts
@@ -1,8 +1,8 @@
 import fs from 'fs'
 import { ethers } from '@nomiclabs/buidler'
-import { bigInt, SnarkBigInt, IncrementalQuinTree } from 'maci-crypto'
 
 import { getEventArg } from '../tests/utils'
+import { getRecipientClaimData } from '../utils/maci'
 
 async function main() {
   const [,,, recipient1, recipient2] = await ethers.getSigners()
@@ -16,34 +16,14 @@ async function main() {
   console.log('Round finalized, totals verified.')
 
   const fundingRound = await ethers.getContractAt('FundingRound', state.fundingRound)
-  const treeDepth =  2
   // Claim funds
   for (const recipientIndex of [1, 2]) {
     const recipient = recipientIndex === 1 ? recipient1 : recipient2
-    const result = parseInt(tally.results.tally[recipientIndex])
-    const resultSalt = tally.results.salt
-    const resultTree = new IncrementalQuinTree(treeDepth, bigInt(0))
-    for (const leaf of tally.results.tally) {
-      resultTree.insert(leaf)
-    }
-    const resultProof = resultTree.genMerklePath(recipientIndex)
-
-    const spent = parseInt(tally.totalVoiceCreditsPerVoteOption.tally[recipientIndex])
-    const spentSalt = tally.totalVoiceCreditsPerVoteOption.salt
-    const spentTree = new IncrementalQuinTree(treeDepth, bigInt(0))
-    for (const leaf of tally.totalVoiceCreditsPerVoteOption.tally) {
-      spentTree.insert(leaf)
-    }
-    const spentProof = spentTree.genMerklePath(recipientIndex)
-    const recipientClaimData = [
-      recipient.getAddress(),
-      result,
-      resultProof.pathElements.map((x) => x.map((y: SnarkBigInt) => y.toString())),
-      resultSalt,
-      spent,
-      spentProof.pathElements.map((x) => x.map((y: SnarkBigInt) => y.toString())),
-      spentSalt,
-    ]
+    const recipientClaimData = getRecipientClaimData(
+      await recipient.getAddress(),
+      recipientIndex,
+      tally,
+    )
     const fundingRoundAsRecipient = fundingRound.connect(recipient)
     const claimTx = await fundingRoundAsRecipient.claimFunds(...recipientClaimData)
     const claimedAmount = await getEventArg(claimTx, fundingRound, 'FundsClaimed', '_amount')
